+++ examples/complete/data.tf
@@ -1,5 +1,3 @@
-data "aws_caller_identity" "current" {}
-
 data "aws_vpc" "supporting" {
   filter {
     name   = "tag:Name"
@@ -22,7 +20,7 @@ data "aws_subnet" "public" {
 data "aws_elb_service_account" "main" {}
 
 ### LB Bucket Policy
-data "aws_iam_policy_document" "alb_s3" {
+data "aws_iam_policy_document" "s3" {
   policy_id = "s3_bucket_lb_logs"
 
   statement {
@@ -31,56 +29,36 @@ data "aws_iam_policy_document" "alb_s3" {
     ]
     effect = "Allow"
     resources = [
-      "arn:aws:s3:::${var.name}-${local.account_id}/*",
+      "arn:aws:s3:::${var.name}/*",
     ]
+
     principals {
       identifiers = [local.service_account]
       type        = "AWS"
     }
   }
-  statement {
-    actions   = ["s3:PutObject"]
-    effect    = "Allow"
-    resources = ["arn:aws:s3:::${var.name}-${local.account_id}/*"]
-    principals {
-      identifiers = ["delivery.logs.amazonaws.com"]
-      type        = "Service"
-    }
-  }
-  statement {
-    actions   = ["s3:GetBucketAcl"]
-    effect    = "Allow"
-    resources = ["arn:aws:s3:::${var.name}-${local.account_id}"]
-    principals {
-      identifiers = ["delivery.logs.amazonaws.com"]
-      type        = "Service"
-    }
-  }
-}
 
-# WAF Bucket Policy
-data "aws_iam_policy_document" "waf_logs_s3" {
   statement {
-    effect = "Allow"
+    actions = [
+      "s3:PutObject"
+    ]
+    effect    = "Allow"
+    resources = ["arn:aws:s3:::${var.name}/*"]
     principals {
-      type        = "Service"
       identifiers = ["delivery.logs.amazonaws.com"]
-    }
-    actions   = ["s3:PutObject"]
-    resources = ["arn:aws:s3:::aws-waf-logs-${var.name}-${local.account_id}/*"]
-    condition {
-      test     = "StringEquals"
-      variable = "s3:x-amz-acl"
-      values   = ["bucket-owner-full-control"]
+      type        = "Service"
     }
   }
+
   statement {
-    effect = "Allow"
+    actions = [
+      "s3:GetBucketAcl"
+    ]
+    effect    = "Allow"
+    resources = ["arn:aws:s3:::${var.name}"]
     principals {
-      type        = "Service"
       identifiers = ["delivery.logs.amazonaws.com"]
+      type        = "Service"
     }
-    actions   = ["s3:GetBucketAcl"]
-    resources = ["arn:aws:s3:::aws-waf-logs-${var.name}-${local.account_id}"]
   }
 }