+++ outputs.tf
@@ -1,4 +1,3 @@
-
 output "arn" {
   description = "The ARN of the WAF WebACL."
   value       = aws_wafv2_web_acl.main.arn